refactor logic to manage test login
